Update mkdocs.yml
I have moved the installation tutorial and the getting started to the top of "For Users". IMO these should be on the top so it feels like more of a guide with progression from no retroarch to installed with all cores and games. If you can understand what I mean.