Added section "Creating custom playlists"
1) Corrected a statement which was plain wrong:
Section "6-Line Playlist Format" explained the 6th entry ("db_name" in the JSON format) as "The name of this playlist, ending in the extension .lpl"
This is simply wrong. The purpose of this entry is rather: Which platform shall I use for looking up metadata and thumbnail or that ROM? So it must be the exact name of the respective platform playlist. For playlists which are platform playlists this is of course their own name. But for cross-platform lists this should always be the respective platform name for metadata/thumbnail lookup to work.

2) Added section "Creating custom playlists"
- Was not explained anywhere. Found out by studying playlist files and doing some trial and error.
- Want to share that knowhow with the community! Should be in the user documentation!
- Created annotated screenshots with great care. Explains the whole concept very visually.

3) Sections "JSON Playlist Format" and "6-Line Playlist Format" have quite some overlap
- There's content redundancy.
- Luckily nothing is contradicting, from what I compared.
- I may clean up that overlap at a later time.
- But for now # 2 was the most important job for me
- In total I took a full working day for this (research, trial'n'error, screenshots, writeup) so it is important that this first changes make it into the user docs.
- Further refinements at a later time hopefully.

4) I plan to commit changes in `roms-playlists-thumbnails.md` and my annotated screenshots in one pass.
But maybe incidentally it will be in two passes, as I use not a real git, but the web interface of GitHub right now.
So you may wait to accept the pull request until the screenshots are included and embedded too.