Update package list for switch-libnx.md
Trying to follow switch-libnx build instructions with an up-to-date devkitpro toolchain resulted in the linker failing to find many libraries. Whereas the Makefile is correct and up-to-date relative to the SDK, this documentation page apparently still reported an old list of required packages, which would be now incomplete. Simply installing 3 additional packages fixes the build issue.