Switch mkdocs from JS theming to Pygments at build
This commit changes several configuration options for MkDocs
that were originally made in an effort to debug the unexpected
rendering of install-gnu.md following commit a746567c. In summary, those are:

- Remove deprecated Python Markdown extension nl2br.
- Switch from client-side syntax highlighting via JavaScript
  to using Pygments at build time to bake the highlighting into
  the generated output. This had the benefit of much better
  determinism and cost only an extra minute added to the build.
- Remove deprecated CodeHilite extension in favor recommended
  Highlight and InlineHilite extensions.

Ultimately, the documentation site loads more quickly and has
greater simplicity, which seemed worth the ~50% increase in
time to generate.

Signed-off-by: Peter J. Mello <admin@petermello.net>