Correct layout issues discovered post-libretro#973
The enhancements to the Content Security Policy and sandboxing
introduced in commit a746567c ultimately prevented the embedded
YouTube video from launching. This commit makes the allow lists
more explicit about safe domains and adds 'allow-same-origin'
and 'allow-scripts' to the sandbox permissions. Also the nested
sublist blocks needed to be indented by multiples of 4 spaces
and not aligned with the first line as needed by other popular
Markdown parsers.

Signed-off-by: Peter J. Mello <admin@petermello.net>