Improve image compression and remove cruft
While investigating other issues it was observed that several
sets of images in the repository were larger than they needed
to be. This commit corrects that oversight with only lossless
methods, yielding just under an average of a 50% reduction in
file size. Also several cruft files such as text editor backups
and .DS_Store files were removed as encountered.

Signed-off-by: Peter J. Mello <admin@petermello.net>