Update generating-retroarch-logs.md to prevent pwsh confusion on Windows
The Windows directions contain a tip to use a folder context menu to open a PowerShell instance in that location. However, the remainder of the commands will fail as pwsh, like bash and other *nix-style shells, does not include . in the path. To sidestep this and avoid having parallel sets of steps (one including "./" and one without), a step was added to the tip to switch to cmd within that PowerShell instance.